nir: Add code to fixup bitfield_insert/extract.
The OpenGL specifications for bitfieldInsert() and bitfieldExtract() says:

   The result will be undefined if <offset> or <bits> is negative, or if
   the sum of <offset> and <bits> is greater than the number of bits
   used to store the operand.

Therefore passing bits=32, offset=0 is legal and defined in GLSL.

But the earlier DX11/SM5 bfi/ibfe/ubfe opcodes are specified to accept a
bitfield width ranging from 0-31. As such, Intel and AMD instructions
read only the low 5 bits of the width operand, making them not able to
implement the GLSL-specified behavior directly.

This commit adds a pass that inserts code to implement the trivial cases
of <bits> = 32 as

   bitfieldInsert:
      bits > 31 ? insert : bitfieldInsert(base, insert, offset, bits)

   bitfieldExtract:
      bits > 31 ? value : bitfieldExtract(value, offset, bits)